Reorganized code to be more readable
