+++ client/src/assets/images/simone.jpg
