+++ client/src/components/ProjectCard.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
 import FaGithub from 'react-icons/lib/fa/github';
 import Chip from 'material-ui/Chip';
-import Colors from './colors.js';
+import Colors from '../assets/colors.js';
 
 class ProjectCard extends Component{
   state={
@@ -40,7 +40,7 @@ class ProjectCard extends Component{
       }
 
       //Establish title and language
-      
+
     }
     catch (e){
       console.log(e);
