+++ client/src/components/SimpleMediaCard.js
