+++ client/src/components/Title.js
@@ -1,11 +1,11 @@
 import React, {Component} from 'react';
 
 import Particles from 'react-particles-js';
-import particles from './particlesjs-config.json';
+import particles from '../assets/particlesjs-config.json';
 import { fadeInUp } from 'react-animations';
 import Radium, { StyleRoot } from 'radium';
-import ACC_Banner from './images/acc_banner.png';
-import './TitleDiv.css';
+import ACC_Banner from '../images/acc_banner.png';
+import '../assets/TitleDiv.css';
 
 const styles = {
   titleText: {
